STEALTH-112 Start of Map Manager - Explore view
Use doubletrouble8082 profile when testing.

Notes:
(1) Only includes a basic impl of Context category.  Futher work under future tickets.
(2) Includes a very rough Static data category. Committing this as-is because it was already in progress.  There is talk of the Timelapse category encompassing all of the Static functionality.  So, we'll revisit after that category is in place.

Reviewed by @perdue