STEALTH-154 Create options wizard for timelapse queries.
Adds a query options wizard for specifying query parameters.
Also executes the query when done with the wizard and adds the resulting
store to the HistoricalLayer.
Notes:
The live layer options and query execution still need to be implemented.
This adds OpenLayers 2 to the javascript libraries to leverage its
WFSGetCapabilities support.
The server url option is only for development purposes and should be
removed in the future.

Reviewed by @rattai