STEALTH-166 Query active stores on map click.
When a user clicks on the map, query the current active stores for
points in a small bounding box around the click location. Currently
only searches the Historical Layer, which determines the bounding
box by its current pixel radius. For file stores the search just
looks through the arraybuffer, while query stores pass a get feature
request back to the server.
The returned data is presented to the user in a table formatted map
overlay.

Reviewed by @rattai