STEALTH-182 Query via map details
Allows stores to specify if map click results are filterable.
If they are then the popup displays them as a link and passes
selected values to a callback function specified by the store.
Also changes where the popup type formatting occurs. To change
formats add a fieldTypes object to the search results that specifies
an object with name and localType properties.

Reviewed by @rattai