STEALTH-120,STEALTH-221,STEALTH-223 Show target history.
The buttons in the expanded target identification results page
will now launch the time-enabled and static data wizards.
The two buttons have also been added to the map popup for the
target identification map points.
Also adds a keyword length check to the querybin store to prevent
an exception when a keyword is shorter than expected.

Reviewed by @rattai