STEALTH-248 Browser suggestions.
Added maxVersion support for selecting a range of selected browser
versions. Added error toasters for unsupported browser and version
and a warning toaster for non-preferred browsers.

Note: Error toasters will not be seen unless we decide to remove
or modify the access block in index.html.

Also, added a closing ` to the readme to fix a code block.

Reviewed by @rattai