STEALTH-277 Live to Historical query capability
Adds a capability to the Live view popup to launch a historical
query. The capability looks for two keywords to help populate default
query params. The first is stealth.capability.timelapse.trkIdField, which
should define the id field to use for creating a CQL filter. The second is
stealth.capability.timelapse.layerName, which specifies the name of the
historical layer to use for the query.
Also fixes an issue where the wrong radio button is selected on the wizard's
time bound step when launching with startDtg and/or endDtg overrides.

Reviewed by @rattai