STEALTH-312 Highlight hooked track
Added a new highlightManager service that can draw highlight diamonds
around point features.
Modified the air live plugin to allow items with open popups to sync
with the layer's polling cycle to refresh the contents of the open popup
and keep the head of the track highlighted. Also changed the air live
popup to show the highlight color of the matching track as a border
in the popup header.

Reviewed by @rattai