STEALTH-313 Follow hooked track.
Adds a follow track button to the live air popup that anchors the center
of the screen on the head of the track. If the popup is closed, the tab
is changed, a different popup's follow button is pressed or the button
in the current popup is pressed again, track following will be cancelled.
Disables map panning while a track is being actively followed.

Reviewed by @rattai