TCRI-130 Users should be able to run route analysis on spatial predictions
The DCM plugin now emits an event to refresh the list of route analysis layers after
adding the new spatial prediction to the map, which is heard by the route analysis plugin. The new
spatial prediction will contain the necessary keywords to appear in the list of
rasters to perform route analysis on.

Reviewed by @rattai