STEALTH-420 Allow Stealth/GeoServer users to save to own workspace, with appropriate roles added
Includes an example spring-security.xml to test example PKI certs with Stealth and have these users
able to save to GeoServer in their own workspaces with roles that lock down the workspace to only that
user having read/write permissions. Also, a small change was made to the DCM plugin to save to a user's
workspace, assuming that user is not anonymous.

Reviewed by @mm64z