STEALTH-300 Core Popup Service
Moves the popup creation and display to a service. Converts the core map search
to use the new service but plugins still use the old way. Moves the default
popup widgets into a core/popup/components directory and refactors the paged table.

Also adds an if check back to the wms popup clear.

Reviewed by @rattai